requirements.txt
streamlit>=1.28.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0
Pillow>=10.0.0
pandas>=2.0.0
python-json-logger>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0